% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.propPA}
\alias{plot.propPA}
\title{Plot proportion of different class of physical activity (PA) across time}
\usage{
\method{plot}{propPA}(object)
}
\arguments{
\item{object}{Output of STEPHEN with steps, HR and predicted_PA as columns.}
}
\value{
A ggplot object
}
\description{
Plot proportion of different class of physical activity (PA) across time
}
