% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_data}
\alias{plot_data}
\title{Plot Step Count and Heart Rate Data coloured by STEPHEN's Physical Activity Class.}
\usage{
plot_data(object)
}
\arguments{
\item{object}{Output of STEPHEN with steps, HR and predicted_PA as columns.}
}
\value{
A ggplot object
}
\description{
Plot Step Count and Heart Rate Data coloured by STEPHEN's Physical Activity Class.
}
